Reorganize Cypress, move test
It's less clutter to put all the cypress-related files into a
subdirectory.  This puts them in %cypress/

This also moves the script %rx-app.cy.js into this repository.
It will be pulled by Ren-C when doing new web builds, and tested
against before greenlighting the build.  But it should be edited
and maintained here.