Rewrite parse rules for new PARSE (UPARSE)
The minimal usage of parsing in this app makes it a good
candidate for using new features.

This doesn't use any of them yet --just changes COPY to
ACROSS and uses the SET-WORD! form of assignment.

Pursuant to the new concept that INTEGER! is treated as itself
and not a repeat count, this also changes NHI-RULE from:

    nhi-rule: [3 alpha 4 digit]

to:

    nhi-rule: [repeat 3 alpha, repeat 4 digit]

Note: PARSE3 is still available--and should be chosen for
code that processes large amounts of data (as UPARSE is still
very slow, being written in usermode, and suboptimal even so).