Adjust skinned RETURN behavior
The system's RETURN is now no longer tolerant of arity 0
situations, and all FUNC must explicitly RETURN.  There are
good reasons for this choice, but people should be able to
skin the system to their tastes.

This makes it so that any function with `[return: <none>]` in
the spec will have an arity-0 RETURN.  It also has a safety
mechanism to ensure the RETURN is at an end and not missing
an argument accidentally.

It also includes the behavior of having functions return
the last value which drops out of their body if no return
is present.