BLANK! no longer falsey, use NULL
Usage of BLANK! is now pretty much exclusively as a placeholder
in blocks.  It is no longer falsey, though it is "EMPTY?"

Most usages for uninitialized fields should use NULL instead.
It has the advantage of not causing errors from word access,
but not being tolerated by most function arguments...so the
uninitializedness is easy to act upon.

NULL can also be easily converted to VOID with MAYBE to opt out
of many operations (the VOID-in-NULL-out protocol, along with
other applications).