Use TRY to defuse out of range pick errors
It used to be that block.3 would just give you null if 3 was out of
range.  This was changed to be an out of range error, that you have
to use TRY to turn into a null.

The issue is still a bit contentious, e.g. that blocks can't have
nulls in them and they're not accepted most places...so it may be
overkill.  Some questions about :block.3 are arising as a possible
way of saying "null if not there".  It's under review, but make
the code here comply for now.