LOAD FILENAME instead of LOAD TO TEXT! FILENAME
There's questions about BLOB! conversions "to text!" in terms of
whether implicitly converting binary to UTF-8 is good or not.  But
the AS command does aliasing of the intrinsic bytes underneath,
which can only be done with UTF-8 given UTF-8 everywhere.

So this could be LOAD AS TEXT! FILENAME, but really LOAD is supposed
to just work on filenames, so do it directly.